+++ src/Errors/Files/FileNotWriteable.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace LBF\Errors\Files;
+
+use LBF\Errors\Meta\ExceptionMeta;
+
+/**
+ * Error page for handling errors when a file is not writeable.
+ * 
+ * use LBF\Errors\Files\FileNotWriteable;
+ * 
+ * @author  Gareth Palmer   [Github & Gitlab /projector22]
+ * @since   LBF 0.1.11-beta
+ */
+
+class FileNotWriteable extends ExceptionMeta {}